Works! 3 random orthog patterns
3 random orthogonal patterns of -1s and 1s, presented twice each in
random order. Test pattern is a randomly chosen input pattern with half
of its bits zero'ed out.

Patterns are 28 bits long, the network has 29 neurons due to inclusion
of a fixed-output neuron that plays the role of a bias.

The solution is very much as expected: essentially zero recurrent fixed
weights (except for distinctly negative self-weights), and recurrent positive
plastic weights.

Interstingly, the fixed-output neuron
seems to project slightly negative plastic projections to everyone.
Presumably this acts a a a damper, and is similar to the negative
self-weights in effect (it reduces the magnitude of correlation with the
fixed output, which means making it less similar to whatever it was
recently). Curiously, the plastic inputs to the fixed-output neuron are
almost exactly zero, with remarkably little variance... and none of them
negative!